[seqio]: Add configurable shuffle_fn function to Task definition.
This exposes a new shuffle_fn property in Task() constructor. It allows callers
to specify a callback function that controls dataset shuffling.  By default,
current users should see no difference as a default shuffling function is
provided with same behavior as before.

This functionality is for users who may want to specify constraints on
shuffling. For example, in our task there are certain examples that should stay
together in the batch (our loss in a batch-wide loss). For this, it becomes
quite simple to define a new shuffling function:

def shuffle_fn(ds, ...):
   ds = ds.batch(num_elements_to_stay_together)
   ds = ds.shuffle(buffer_size)
   ds = ds.unbatch()
   return ds

PiperOrigin-RevId: 386561048