[seqio]: Expose dataset sampling function in Mixtures, allowing users to control how datasets are interleaved.
This adds a new `sample_fn` property to Mixture() which controls how datasets
are interleaved (sampled). Currently, the
`tf.data.experimental.sample_from_datasets` is used for sampling, and the
default behavior is preserved for current users.

I expect this may be useful for tasks where there is some data or task dependent
constraint on how the sequence (or batch) of elements should be placed. One
simple example are tasks that contain batch-level losses that require certain
task elements to be placed in sequence in the batch.

PiperOrigin-RevId: 388448841