Allow DataSources to come bundled with evaluation metrics.
This is to support the concept of a standard benchmark, consisting of both a test dataset and metrics to be computed.

One approach might have been to add a `benchmark` argument to the Task constructor, accepting a type that bundles (DataSource, metric_fns); but that would have required making the existing `source` argument optional.  Here, instead, we extend the DataSource type, as this seems less disruptive to the Task constructor signature (and to TaskRegistry.add(), etc.) and generally more expedient.

PiperOrigin-RevId: 395979787