Add a charater count in the caching job.
Currently, we only compute the number of tokens for tokenized datasets. Since
the number of characters in each token can vary depending on the task, it can
be useful to also compute the number of characters in each example.

* Concrete use case:
For multilingual datasets. The average char len of each language's token can
be very different. So the token counts or example count are not good proxies
for representing the "size" of each language, which is in turn used to
determine the "sampling rate" of the language in the mixture. Character count
is a better statistic.

* Added behavior
This change adds `_CountCharacters` beam `DoFn` that are applied to the
examples PCollection. For tokenized datasets, the task's vocabulary is used to
detokenized the data and computes the string length. For string datasets, the
bytes data is directly decoded.

Test:
Both tokenized and string datasets are cached and their statistics were
verified by manually computing the character length in colab.
PiperOrigin-RevId: 451198932