[seqio] Extends `test_utils.assert_dataset()` to work with bare float values inside a dict.
Prior to this CL, `assert_dataset()` works with examples where some of the values inside the dictionary are `ndarrays` of floats, but it fails if some of the values inside the dictionary are raw floats. That is, it raises an assertion error when in fact the two datasets being compared are equal.

After this CL, `assert_dataset()` also works with bare float features inside the dictionary.

The new assertions added to the unit test file `test_utils_test.py` fail for the code in `test_utils.py` in HEAD, but succeed with this CL.

PiperOrigin-RevId: 463895330