Fixes bug introduced in Task.replace
A few attributes of Task were made lazy to help with slow startup times. Task.replace needed to be updated, but was not.

This was not detected because there was no test coverage.

This CL includes a lot more tests for Task.replace.

It also fixes a long-open bug where Task.replace silently swallowed args that were not part of the Task constructor.

PiperOrigin-RevId: 518683483