Update seqio.preprocessors.print_dataset for TF2
`tf.Print()` is gone. `tf.print()` is the successor, however it doesn't work in this context anymore, at least when this preprocessor is the last one in the chain, because the return value needs to be compatible with TypeSpec and so we get `Could not build a TypeSpec for name: "PrintV2"`. The [solution](https://stackoverflow.com/questions/70007364/tf-print-gives-typespec-typeerror) is to use a function like `tf.keras.backend.print_tensor()` (and why not use that one directly here), which internally uses `tf.control_dependencies()` and returns a `tf.identity()` instead, such that the TypeSpec still works.

PiperOrigin-RevId: 580211423